ImageProcess directory moved to Image
